+++ source/core/grids.cc
@@ -390,9 +390,9 @@ read_mesh_and_manifolds(
         {
           double minimal_cell_size =
             GridTools::minimal_cell_diameter(triangulation);
-          double       target_size = mesh_parameters.target_size;
-          unsigned int number_refinement =
-            floor(std::log(minimal_cell_size / target_size) / std::log(2));
+          double       target_size       = mesh_parameters.target_size;
+          unsigned int number_refinement = static_cast<unsigned int>(
+            floor(std::log(minimal_cell_size / target_size) / std::log(2)));
           triangulation.refine_global(number_refinement);
         }
       else if (!restart)
@@ -613,9 +613,9 @@ read_mesh_and_manifolds_for_stator_and_rotor(
         {
           const double minimal_cell_size =
             GridTools::minimal_cell_diameter(triangulation);
-          const double       target_size = mesh_parameters.target_size;
-          const unsigned int number_refinement =
-            floor(std::log(minimal_cell_size / target_size) / std::log(2));
+          const double       target_size       = mesh_parameters.target_size;
+          const unsigned int number_refinement = static_cast<unsigned int>(
+            floor(std::log(minimal_cell_size / target_size) / std::log(2)));
           triangulation.refine_global(number_refinement);
         }
       else if (!restart)
